+++ package.json
@@ -44,7 +44,7 @@
     "build": "npm run build-to5 && npm run build-browser && npm run build-minify",
     "build-browser": "browserify ./lib/index.js --standalone WavEncoder --outfile build/wav-encoder.js",
     "build-minify": "uglifyjs build/wav-encoder.js -m -r self -o build/wav-encoder.min.js",
-    "build-to5": "babel src --out-dir lib && cp src/encoder-worker.js lib/",
+    "build-to5": "babel src --out-dir lib && cp src/EncoderWorker.js lib/",
     "cover": "babel-node $(npm bin)/isparta cover --report text --report html --report lcov _mocha",
     "lint": "eslint src test",
     "test": "mocha --compilers js:espower-babel/guess",
