+++ src/EncoderWorker.js
